rename mock.py in core
