+++ src/azure-cli-core/azure/cli/core/mock_cli.py
