+++ src/azure-cli-core/azure/cli/core/tests/test_telemetry.py
@@ -43,7 +43,7 @@ class TestCoreTelemetry(unittest.TestCase):
     def test_cloud_forbid_telemetry(self):
         import mock
         import azure.cli.core.telemetry as telemetry
-        from azure.cli.core.mock import DummyCli
+        from azure.cli.core.mock_cli import DummyCli
         from knack.completion import ARGCOMPLETE_ENV_NAME
 
         az_cli = DummyCli()
