+++ src/azure-cli/azure/cli/command_modules/batch/tests/latest/batch_preparers.py
@@ -14,7 +14,7 @@ class BatchAccountPreparer(AbstractPreparer, SingleValueReplacer):
                  resource_group_parameter_name='resource_group', skip_delete=True,
                  dev_setting_name='AZURE_CLI_TEST_DEV_BATCH_ACCT_NAME'):
         super(BatchAccountPreparer, self).__init__(name_prefix, 24)
-        from azure.cli.core.mock import DummyCli
+        from azure.cli.core.mock_cli import DummyCli
         self.cli_ctx = DummyCli()
         self.parameter_name = parameter_name
         self.resource_group = None
