+++ src/azure-cli/azure/cli/command_modules/profile/tests/latest/test_profile_custom.py
@@ -7,7 +7,7 @@ import unittest
 import mock
 
 from azure.cli.command_modules.profile.custom import list_subscriptions, get_access_token, login
-from azure.cli.core.mock import DummyCli
+from azure.cli.core.mock_cli import DummyCli
 
 
 class ProfileCommandTest(unittest.TestCase):
