+++ src/azure-cli/azure/cli/command_modules/resource/tests/hybrid_2019_03_01/test_api_check.py
@@ -51,7 +51,7 @@ class TestApiCheck(unittest.TestCase):
 
     def test_resolve_api_provider_backup(self):
         # Verifies provider is used as backup if api-version not specified.
-        from azure.cli.core.mock import DummyCli
+        from azure.cli.core.mock_cli import DummyCli
         cli = DummyCli()
         rcf = self._get_mock_client()
         res_utils = _ResourceUtils(cli, resource_type='Mock/test', resource_name='vnet1',
@@ -60,7 +60,7 @@ class TestApiCheck(unittest.TestCase):
 
     def test_resolve_api_provider_with_parent_backup(self):
         # Verifies provider (with parent) is used as backup if api-version not specified.
-        from azure.cli.core.mock import DummyCli
+        from azure.cli.core.mock_cli import DummyCli
         cli = DummyCli()
         rcf = self._get_mock_client()
         res_utils = _ResourceUtils(cli, parent_resource_path='foo/testfoo123', resource_group_name='rg',
@@ -71,7 +71,7 @@ class TestApiCheck(unittest.TestCase):
 
     def test_resolve_api_all_previews(self):
         # Verifies most recent preview version returned only if there are no non-preview versions.
-        from azure.cli.core.mock import DummyCli
+        from azure.cli.core.mock_cli import DummyCli
         cli = DummyCli()
         rcf = self._get_mock_client()
         res_utils = _ResourceUtils(cli, resource_type='Mock/preview', resource_name='vnet1',
