+++ src/azure-cli/azure/cli/command_modules/vm/tests/hybrid_2018_03_01/test_vm_defaults.py
@@ -25,7 +25,7 @@ from azure.cli.command_modules.vm._validators import (_validate_vm_vmss_create_v
 
 
 def _get_test_cmd():
-    from azure.cli.core.mock import DummyCli
+    from azure.cli.core.mock_cli import DummyCli
     from azure.cli.core import AzCommandsLoader
     from azure.cli.core.commands import AzCliCommand
     cli_ctx = DummyCli()
