+++ src/azure-cli/azure/cli/command_modules/vm/tests/latest/test_vm_actions.py
@@ -22,7 +22,7 @@ from azure.cli.command_modules.vm._validators import (validate_ssh_key,
                                                       _validate_vm_vmss_accelerated_networking,
                                                       process_gallery_image_version_namespace)
 from azure.cli.command_modules.vm._vm_utils import normalize_disk_info, update_disk_sku_info
-from azure.cli.core.mock import DummyCli
+from azure.cli.core.mock_cli import DummyCli
 from knack.util import CLIError
